refactor: part rendering moves out of CityGenerator (#11)
Parts holds the block-layout loop and everything a palette entry can
attach to a block: block entities and their NBT, spawners, loot tables,
and the deferred todos that need a real world rather than a half-built
chunk. The block-entity type cache moves with getTypeForBlock, which is
its only reader.

In `worldgen` rather than `worldgen.gen`, unlike the other passes split
out so far. This one queues deferred work through
ChunkGenContext.addPostTodo and addLightTodo, which are package-private
on purpose: issue #127 moved runtime callbacks off the planning types
precisely so that queueing them stayed inside this package. Making them
public to let this class sit beside its siblings would undo that, so the
class moved instead of the boundary.

hardAir joins air and liquid as public - the same three block states, and
generatePart is the reason the other two were already public.

Pure code motion: 1929 tokens on both sides, the only difference being
`this.provider` becoming a parameter reference.

CityGenerator 1420 -> 1030 lines; 2425 at the start of this milestone.

All seven digest suites unchanged.
