+++ src/main/java/dev/krona/urbex/worldgen/gen/Highways.java
@@ -3,6 +3,7 @@ package dev.krona.urbex.worldgen.gen;
 import dev.krona.urbex.worldgen.ChunkDriver;
 import dev.krona.urbex.worldgen.ChunkGenContext;
 import dev.krona.urbex.worldgen.CityGenerator;
+import dev.krona.urbex.worldgen.Parts;
 import dev.krona.urbex.worldgen.lost.ChunkPlan;
 import dev.krona.urbex.worldgen.lost.Highway;
 import dev.krona.urbex.worldgen.lost.Transform;
@@ -50,12 +51,12 @@ public class Highways {
         if (info.isTunnel(level)) {
             // We know we need a tunnel
             part = info.provider.assets().parts().getOrThrow(feature.getRandomPart(ctx, highwayParts.tunnel(bidirectional)));
-            feature.generatePart(ctx, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
+            Parts.generatePart(ctx, feature, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
         } else {
             if (info.isCity && level <= adjacent1.cityLevel && level <= adjacent2.cityLevel && adjacent1.isCity && adjacent2.isCity) {
                 // Simple highway in the city
                 part = info.provider.assets().parts().getOrThrow(feature.getRandomPart(ctx, highwayParts.open(bidirectional)));
-                int height = feature.generatePart(ctx, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
+                int height = Parts.generatePart(ctx, feature, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
                 // Clear a bit more above the highway
                 if (!info.profile.isCavern()) {
                     int clearheight = 15;
@@ -68,7 +69,7 @@ public class Highways {
                 }
             } else {
                 part = info.provider.assets().parts().getOrThrow(feature.getRandomPart(ctx, highwayParts.bridge(bidirectional)));
-                int height = feature.generatePart(ctx, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
+                int height = Parts.generatePart(ctx, feature, info, part, transform, 0, highwayGroundLevel, 0, CityGenerator.HardAirSetting.WATERLEVEL);
                 // Clear a bit more above the highway
                 if (!info.profile.isCavern()) {
                     int clearheight = 15;
